Added an easier way to navigate between the user and admin sections.
Through .ejs templating I was able to figure how to link to the lecture and chapter pages after a submission on the admin side. On the user side, links to the admin section pop up when the user is also an admin.

These changes can be seen by going to the lecture page and clicking on a chapteror by going to the backend and editing something.